backend/requirements/base.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
drf-spectacular==0.27.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.1
django-oauth-toolkit==2.4.0
cryptography==42.0.0
argon2-cffi==23.1.0  # Secure password hashing


# Async & Background Tasks
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
kombu==5.3.5

# File Management
Pillow>=11.0.0
django-storages==1.14.2
boto3==1.34.34

# Utilities
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.1
python-dateutil==2.8.2

# Excel & CSV
openpyxl==3.1.2
xlsxwriter==3.2.0
pandas==2.2.0

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# API & HTTP
requests==2.31.0
urllib3==2.2.0

# Validation
django-phonenumber-field==7.3.0
phonenumbers==8.13.29

# Monitoring & Logging
sentry-sdk==1.40.0

# Code Quality
django-extensions==3.2.3
